Cleaning up .gitignore and reorganize
