+++ Bloggity/settings.py
@@ -12,7 +12,7 @@ https://docs.djangoproject.com/en/4.2/ref/settings/
 from datetime import timedelta
 from pathlib import Path
 
-from utils.secrets_utils import GoogleCloudsSecretManager
+from GoogleClouds.secrets_utils import GoogleCloudsSecretManager
 
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 BASE_DIR = Path(__file__).resolve().parent.parent
