+++ Dockerfile
@@ -11,20 +11,8 @@ ENV PYTHONDONTWRITEBYTECODE=1
 # the application crashes without emitting any logs due to buffering.
 ENV PYTHONUNBUFFERED=1
 
-#TODO: Assuming local environment, will remove this when connecting to Google clouds.
-ENV DEBUG=True
-ENV PATH_TO_DJANGO_SETTINGS='Bloggity.settings.local'
-ENV DJANGO_SECRET_KEY="django-insecure-kih$vse%bf+e9%4=ii7yye+s120^r8ug!5$4@k@3hnfsk+@i%r"
-ENV DB_NAME="Bloggity"
-ENV DB_USER="postgres"
-ENV DB_PASS="12345"
-ENV DB_HOST="db"
-ENV DB_PORT=5432
-ENV STATIC_URL="static/"
-ENV ALLOWED_HOSTS="localhost,127.0.0.1"
-
 # Installing Python
-RUN apk update && apk add --no-cache python3=3.11.9-r0 \ 
+RUN apk update && apk add --no-cache python3=3.11.9-r0 \
 && apk add --no-cache py3-pip && python3 -m venv /venv && \
 # Installing dependencies for Postgres
 apk add --no-cache libpq-dev && \
